requirements.txt
langchain==0.2.1
langchain_community==0.2.1
tiktoken==0.7.0
langchain-openai==0.1.8 
langchainhub==0.1.17
chromadb==0.5.0
unstructured==0.14.3
camelot-py[cv]
ghostscript==0.7
jq==1.7.0